requirements.txt
playwright
atproto
python-dotenv
requests